-Wall cleanups, continued
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@76961 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/http_protocol.c
@@ -227,7 +227,7 @@ int set_byterange (request_rec *r)
 	
 	r->byterange = 2;
 	table_unset(r->headers_out, "Content-Length");
-	sprintf(boundary, "%lx%lx", r->request_time, getpid());
+	sprintf(boundary, "%lx%lx", r->request_time, (long)getpid());
 	r->boundary = pstrdup(r->pool, boundary);
     }
